👷  (semantic) NICE-67 test build [skip ci]
